requirements.txt
langchain==0.1.13
langchain-openai==0.1.1
openai==1.14.3
langchain-core==0.1.28
langchain-community==0.0.24
psycopg2-binary==2.9.9
psycopg==3.1.18
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
streamlit==1.32.2